Add IPv6.toFixedLengthString.
There are a few different ways to normalize x:0:y:0:z, and this has
historically been a source of bugs. In fact, our implementation
does not quite follow RFC5952. In cases like a database primary key,
it is nicer to use a larger but clearly unambiguous representation,
which is what's implemented here.